+++ packages/expo-font/src/FontLoader.web.ts
@@ -9,7 +9,7 @@ function uriFromFontSource(asset: any): string | null {
   if (typeof asset === 'string') {
     return asset || null;
   } else if (typeof asset === 'object') {
-    return asset.uri || asset.localUri || null;
+    return asset.uri || asset.localUri || asset.default || null;
   }
   return null;
 }