+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/ModuleRegistry.kt
@@ -25,6 +25,7 @@ class ModuleRegistry(
       )
     }
     holder.post(EventName.MODULE_CREATE)
+    holder.registerContracts()
     registry[holder.name] = holder
   }
 
@@ -74,5 +75,7 @@ class ModuleRegistry(
     forEach {
       it.cleanUp()
     }
+    registry.clear()
+    logger.info("✅ ModuleRegistry was destroyed")
   }
 }