+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/CppType.kt
@@ -1,5 +1,10 @@
 package expo.modules.kotlin.jni
 
+import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
+import expo.modules.kotlin.typedarray.TypedArray
+import kotlin.reflect.KClass
+
 private var nextValue = 0
 
 private fun nextValue(): Int = (1 shl nextValue).also { nextValue++ }
@@ -7,19 +12,19 @@ private fun nextValue(): Int = (1 shl nextValue).also { nextValue++ }
 /**
  * Enum that represents Cpp types. Objects of those types can be obtained via JNI.
  */
-enum class CppType(val value: Int = nextValue()) {
-  NONE(0),
-  DOUBLE,
-  INT,
-  FLOAT,
-  BOOLEAN,
-  STRING,
-  JS_OBJECT,
-  JS_VALUE,
-  READABLE_ARRAY,
-  READABLE_MAP,
-  TYPED_ARRAY,
-  PRIMITIVE_ARRAY,
-  LIST,
-  MAP;
+enum class CppType(val clazz: KClass<*>, val value: Int = nextValue()) {
+  NONE(Nothing::class, 0),
+  DOUBLE(Double::class),
+  INT(Int::class),
+  FLOAT(Float::class),
+  BOOLEAN(Boolean::class),
+  STRING(String::class),
+  JS_OBJECT(JavaScriptObject::class),
+  JS_VALUE(JavaScriptValue::class),
+  READABLE_ARRAY(ReadableArray::class),
+  READABLE_MAP(ReadableMap::class),
+  TYPED_ARRAY(TypedArray::class),
+  PRIMITIVE_ARRAY(Array::class),
+  LIST(List::class),
+  MAP(Map::class);
 }