+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/TypeConverterProvider.kt
@@ -3,9 +3,11 @@
 package expo.modules.kotlin.types
 
 import android.graphics.Color
+import android.net.Uri
 import com.facebook.react.bridge.Dynamic
 import com.facebook.react.bridge.ReadableArray
 import com.facebook.react.bridge.ReadableMap
+import expo.modules.kotlin.apifeatures.EitherType
 import expo.modules.kotlin.exception.MissingTypeConverter
 import expo.modules.kotlin.jni.CppType
 import expo.modules.kotlin.jni.ExpectedType
@@ -25,6 +27,15 @@ import expo.modules.kotlin.typedarray.Uint16Array
 import expo.modules.kotlin.typedarray.Uint32Array
 import expo.modules.kotlin.typedarray.Uint8Array
 import expo.modules.kotlin.typedarray.Uint8ClampedArray
+import expo.modules.kotlin.types.io.FileTypeConverter
+import expo.modules.kotlin.types.io.PathTypeConverter
+import expo.modules.kotlin.types.net.JavaURITypeConverter
+import expo.modules.kotlin.types.net.URLTypConverter
+import expo.modules.kotlin.types.net.UriTypeConverter
+import java.io.File
+import java.net.URI
+import java.net.URL
+import java.nio.file.Path
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.reflect.full.createType
@@ -102,7 +113,22 @@ object TypeConverterProviderImpl : TypeConverterProvider {
       return converter
     }
 
-    throw MissingTypeConverter(type)
+    return handelEither(type, kClass) ?: throw MissingTypeConverter(type)
+  }
+
+  @OptIn(EitherType::class)
+  private fun handelEither(type: KType, kClass: KClass<*>): TypeConverter<*>? {
+    if (kClass.isSubclassOf(Either::class)) {
+      if (kClass.isSubclassOf(EitherOfFour::class)) {
+        return EitherOfFourTypeConverter<Any, Any, Any, Any>(this, type)
+      }
+      if (kClass.isSubclassOf(EitherOfThree::class)) {
+        return EitherOfThreeTypeConverter<Any, Any, Any>(this, type)
+      }
+      return EitherTypeConverter<Any, Any>(this, type)
+    }
+
+    return null
   }
 
   private fun createCashedConverters(isOptional: Boolean): Map<KType, TypeConverter<*>> {
@@ -197,6 +223,13 @@ object TypeConverterProviderImpl : TypeConverterProvider {
       TypedArray::class.createType(nullable = isOptional) to TypedArrayTypeConverter(isOptional),
 
       Color::class.createType(nullable = isOptional) to ColorTypeConverter(isOptional),
+
+      URL::class.createType(nullable = isOptional) to URLTypConverter(isOptional),
+      Uri::class.createType(nullable = isOptional) to UriTypeConverter(isOptional),
+      URI::class.createType(nullable = isOptional) to JavaURITypeConverter(isOptional),
+
+      File::class.createType(nullable = isOptional) to FileTypeConverter(isOptional),
+      Path::class.createType(nullable = isOptional) to PathTypeConverter(isOptional),
 
       Any::class.createType(nullable = isOptional) to AnyTypeConverter(isOptional),
     )